Move existing src/dst types in connector into subpackages
