+++ connector/src/main/scala/quasar/connector/datasource/PhysicalDatasource.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quasar.connector
+package quasar.connector.datasource
 
 import quasar.api.resource.ResourcePathType
