+++ impl/src/main/scala/quasar/impl/datasource/ConditionReportingDatasource.scala
@@ -20,7 +20,7 @@ import slamdata.Predef.{Boolean, None, Option, Some, Unit}
 import quasar.Condition
 import quasar.api.datasource.DatasourceType
 import quasar.api.resource._
-import quasar.connector.Datasource
+import quasar.connector.datasource.Datasource
 import quasar.contrib.scalaz.MonadError_
 
 import scalaz.Monad
