+++ impl/src/main/scala/quasar/impl/datasource/FailedDatasource.scala
@@ -19,7 +19,7 @@ package quasar.impl.datasource
 import slamdata.Predef.{Boolean, Option}
 import quasar.api.datasource.DatasourceType
 import quasar.api.resource._
-import quasar.connector.Datasource
+import quasar.connector.datasource.Datasource
 import quasar.contrib.scalaz.MonadError_
 
 import scalaz.Applicative
