+++ impl/src/main/scala/quasar/impl/datasource/local/LocalParsedDatasourceModule.scala
@@ -24,7 +24,8 @@ import quasar.api.datasource.DatasourceError.{
 }
 import quasar.common.data.RValue
 import quasar.{concurrent => qc}
-import quasar.connector._, LightweightDatasourceModule.DS
+import quasar.connector._
+import quasar.connector.datasource.LightweightDatasourceModule
 
 import scala.concurrent.ExecutionContext
 import scala.util.Either
@@ -47,7 +48,7 @@ object LocalParsedDatasourceModule extends LightweightDatasourceModule with Loca
       rateLimiting: RateLimiting[F, A],
       stateStore: ByteStore[F])(
       implicit ec: ExecutionContext)
-      : Resource[F, Either[InitializationError[Json], DS[F]]] = {
+      : Resource[F, Either[InitializationError[Json], LightweightDatasourceModule.DS[F]]] = {
     val ds = for {
       lc <- attemptConfig[F, LocalConfig, InitializationError[Json]](
         config,
