+++ impl/src/main/scala/quasar/impl/datasource/local/LocalStatefulDatasource.scala
@@ -17,8 +17,8 @@
 package quasar.impl.datasource.local
 
 import slamdata.Predef._
-import quasar.connector.LightweightDatasourceModule.DS
 import quasar.connector._
+import quasar.connector.datasource.LightweightDatasourceModule
 
 import java.nio.file.{Path => JPath}
 
@@ -35,7 +35,7 @@ object LocalStatefulDatasource {
       format: DataFormat,
       pageSize: Long,
       blocker: Blocker)
-      : DS[F] = {
+      : LightweightDatasourceModule.DS[F] = {
 
     EvaluableLocalDatasource[F](LocalType, root) { iRead =>
