+++ impl/src/main/scala/quasar/impl/datasources/ManagedDatasource.scala
@@ -19,7 +19,7 @@ package quasar.impl.datasources
 import slamdata.Predef._
 import quasar.api.datasource.DatasourceType
 import quasar.api.resource._
-import quasar.connector.Datasource
+import quasar.connector.datasource.Datasource
 import quasar.qscript.{InterpretedRead, QScriptEducated}
 
 import scalaz.~>
