+++ impl/src/main/scala/quasar/impl/destinations/DestinationModules.scala
@@ -20,7 +20,8 @@ import slamdata.Predef._
 
 import quasar.api.destination._
 import quasar.api.destination.DestinationError._
-import quasar.connector.{DestinationModule, MonadResourceErr}
+import quasar.connector.MonadResourceErr
+import quasar.connector.destination.DestinationModule
 import quasar.impl.IncompatibleModuleException.linkDestination
 
 import argonaut.Json
